Revert "Move BaseResource to types (#4334)"
This reverts commit 6e4b3a5db41651ae876f6fc5cb31991f365e95a9.
