+++ types/src/front/lib/core_api.ts
@@ -8,6 +8,7 @@ import {
   CoreAPILightDocument,
 } from "../../core/data_source";
 import { dustManagedCredentials } from "../../front/lib/api/credentials";
+import { Err, Ok, Result } from "../../front/lib/result";
 import { Project } from "../../front/project";
 import { CredentialsType } from "../../front/provider";
 import {
@@ -18,7 +19,6 @@ import {
   TraceType,
 } from "../../front/run";
 import { LoggerInterface } from "../../shared/logger";
-import { Err, Ok, Result } from "../../shared/result";
 
 const { CORE_API = "http://127.0.0.1:3001" } = process.env;
