+++ types/src/front/lib/result.ts
