+++ types/src/index.ts
@@ -41,6 +41,7 @@ export * from "./front/lib/diff";
 export * from "./front/lib/dust_api";
 export * from "./front/lib/error";
 export * from "./front/lib/extract_event_app";
+export * from "./front/lib/result";
 export * from "./front/membership_invitation";
 export * from "./front/plan";
 export * from "./front/project";
@@ -53,8 +54,6 @@ export * from "./shared/message_classification";
 export * from "./shared/model_id";
 export * from "./shared/nango_errors";
 export * from "./shared/rate_limiter";
-export * from "./shared/resources/base_resource";
-export * from "./shared/result";
 export * from "./shared/user_operation";
 export * from "./shared/utils/assert_never";
 export * from "./shared/utils/config";
